Fix: SC2115: Use \"${var:?}\" to ensure this never expands to / .

+++ test/test_run_examples.sh
@@ -19,7 +19,7 @@ exit_code=0
 for dir in ${program_dirs}
 do
     echo "> cleaning: rm -rf ${BASEDIR}/${dir}"
-    rm -rf ${BASEDIR}/${dir}
+    rm -rf ${BASEDIR:?}/${dir}
 
     echo "> copying: cp -r ${examples_dir}/${dir}/ ${BASEDIR}"
     cp -r ${examples_dir}/${dir}/ ${BASEDIR}
@@ -61,7 +61,7 @@ do
     fi
 
     echo "> cleaning: rm -rf ${BASEDIR}/${dir}"
-    rm -rf ${BASEDIR}/${dir}/
+    rm -rf ${BASEDIR:?}/${dir}/
 done
 
 exit ${exit_code}
